Update SentimentAnalysis.sql
Updated int to be a float

+++ db/SentimentAnalysis.sql
@@ -3,5 +3,5 @@ CREATE TABLE SentimentAnalysis (
     TimeStamp timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
     Text varchar(1000),
     SentimentScore varchar(255),
-    SentimentPercentage int
+    SentimentPercentage float(4)
 );
